// packages/old/src/types/types.ts
export enum Availability {
  AVAILABLE = 'AVAILABLE',
  NOT_APPLICABLE = 'NOT_APPLICABLE',
  NOT_AVAILABLE = 'NOT_AVAILABLE',
  UNKNOWN = 'UNKNOWN',
}
// ...
export enum APIType {
  AP = 'AP',
  EL = 'EL',
  EU = 'EU',
  HK = 'HK',
  UNKNOWN = 'UNKNOWN',
  US = 'US',
}